% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_sd.R
\name{check_sd}
\alias{check_sd}
\title{Check gene-wise STD of each one matrix}
\usage{
check_sd(objt1, objt2, genes, min.sd = 0.001)
}
\arguments{
\item{objt1}{dataset for condition 1, genes in columns}

\item{objt2}{dataset for condition 2, genes in columns}

\item{genes}{gene names or colnames of objt1/objt2}

\item{min.sd}{a valid data matrix per group must have at least this much per-feature STD}
}
\value{
list of retained genes and to-be-removed genes for STD reason
}
\description{
Check if data matrix per group has a minimum or higher per-gene STD
}
\examples{
data(meta)
STAD <- datasets[['STAD']]
N <- ncol(STAD)
n1 <- N\%/\%2
objt1 <- t(STAD[,1:n1])
objt2 <- t(STAD[,(n1+1):N])
genes <- rownames(STAD)
check.res <- check_sd(objt1,objt2,genes,0.1)

}
