% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GSAR_boot.R
\name{GSAR_boot}
\alias{GSAR_boot}
\title{Perform differential coexpression analysis on original and bootstrap datasets}
\usage{
GSAR_boot(
  R,
  gsets,
  object,
  group,
  nperm = 100,
  cor.method = "pearson",
  max.skip = 50,
  min.sd = 0.001,
  minGsize = 3
)
}
\arguments{
\item{R}{number of bootstrap times.}

\item{gsets}{list of multiple gene sets.}

\item{object}{gene expression matrix covering two groups. Row names are gene symbols.}

\item{group}{original groupping of samples, vector of 1's and 2's.}

\item{nperm}{times of sample indix permutation, necessitated by GSNCA}

\item{cor.method}{correlation method}

\item{max.skip}{maximum number of repeated permutation/bootstrap times to avoid zero STD}

\item{min.sd}{a valid data matrix per group must have at least this much per-feature STD}

\item{minGsize}{considered gene set must have this minimum size after overlaying with gene expression matrix.}
}
\value{
list of GSARboot_format() results for multiple gene sets. Component has multiple elements plus a dsetRow which cover all ultimate output
}
\description{
Differential coexpression analysis is conducted on one original dataset and multiple bootstrap datasets.
  
1. Genes/rows in original data failing STD check are simply dropped. Genes/rows in bootstrap data failing STD check are simply dropped, we did not bother trying a second bootstrap for presence of certain minimal STD genes. 2. When original dataset does not contain sufficient genes for running GSNCA on one particular gene set, warning will be shown as GSET: Results NOT Captured!!! The output list simply does not contain component for that GSET.
}
