% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GSARboot_format.R
\name{GSARboot_format}
\alias{GSARboot_format}
\title{Assemble GSNACA results on original dataset and bootstrap datasets}
\usage{
GSARboot_format(
  D_obs,
  pvalue,
  D_obs_boot,
  pvalue_boot,
  genes.not.found = "-",
  genes.removed = "-",
  nperm = 100,
  R = 200,
  level = 0.05,
  call = "Command used"
)
}
\arguments{
\item{D_obs}{distance statistics from original dataset, by GSNCA}

\item{pvalue}{p-value from original dataset, by GSNCA}

\item{D_obs_boot}{distance statistics from multiple bootstrap datasets}

\item{pvalue_boot}{p-values from multiple bootstrap datasets}

\item{genes.not.found}{string of excluded genes for absence in expression matrix}

\item{genes.removed}{string of genes that were removed for STD disqualification}

\item{nperm}{number of permutations passed on to GSNCA}

\item{R}{number of bootstrapping}

\item{level}{confidence interval of this level. Defaults to 0.05 (95% CI)}

\item{call}{external ultimate command exertion statement for GSAR_boot(). A character string}
}
\value{
list of multiple objects of statistics value, where component dsetRow consists of 14 elements to appear as a row in output file
}
\description{
Collate statistics and p-value of GSNCA on original dataset and counterpart statistics or median and CIs from bootstrap datasets
}
\details{
Only one gene set is considered in the current function
}
