% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsnca.stat.R
\name{gsnca.stat}
\alias{gsnca.stat}
\title{Test statistics of Gene Sets Net Correlation Analysis (GSNCA)}
\usage{
gsnca.stat(objt1, objt2, cor.method = "pearson")
}
\arguments{
\item{objt1}{dataset for condition 1, genes in columns}

\item{objt2}{dataset for condition 2, genes in columns}
}
\value{
L1-norm distance between two weight vectors
}
\description{
Derive distance statistics by solving eigenvector of two pairise correaltion matrix.
}
\examples{
data(BRCA)
N <- ncol(BRCA)
n1 <- floor(N/2)
objt1 <- t(BRCA[1:100,1:n1])
objt2 <- t(BRCA[1:100,(n1+1):N])
gStat.res <- gsnca.stat(objt1,objt2)

}
\seealso{
[gsnca.p()] for the external function that calls on the current function and returns permutation p-statistic.
}
