% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsnca_gsets.R
\name{gsnca_gsets}
\alias{gsnca_gsets}
\title{Run GSNCA over multiple gene sets}
\usage{
gsnca_gsets(
  gsets,
  object,
  group,
  perm.list,
  cor.method = "pearson",
  max.skip = 50,
  min.sd = 0.001,
  minGsize = 3
)
}
\arguments{
\item{gsets}{list for multiple gene sets.}

\item{object}{gene expression matrix covering two groups. Row names are gene symbols.}

\item{group}{original groupping of samples, vector of 1's and 2's.}

\item{perm.list}{list of permutation specs. Each component gives permutated sample indices}

\item{cor.method}{correlation method}

\item{max.skip}{maximum number of repeated permutation/bootstrap times to avoid zero STD}

\item{min.sd}{a valid data matrix per group must have at least this much per-feature STD}

\item{minGsize}{considered gene set must have this minimum size after overlaying with gene expression matrix.}
}
\value{
geneset-wise GSNCA results, each consisting of p-value and statistics out of GSNCA.
}
\description{
For each gene-set loop, genes are overlaid to expression matrix and too small gene sets are declined for GSNCA run.
}
\details{
Due to too small gene set size (with consideration of intersection with expression data), certain gene sets have NA as p and stat results.
}
\examples{
data(meta)
BRCA <- datasets[['BRCA']]
smpCode <- substr(colnames(BRCA),14,15)
grp1 <- which(smpCode=='01')
grp2 <- which(smpCode=='11')
object <- BRCA[1:100,c(grp1,grp2)]
group <- c(rep(1,length(grp1)),rep(2,length(grp1)))
perm.list <- vector('list',500)
for (i in seq_len(500)) {perm.list[[i]] <- sample(ncol(object))}
gsets <- split(rownames(object),rep(1:4,each=25))
res <- gsnca_gsets(gsets,object,group,perm.list)

}
\seealso{
[gsnca_p()] for the GSNCA algorithm, which further calls on [gsnca_stat()] for coexpression distance statistics.
}
