% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaAndPlot.R
\name{metaAndPlot}
\alias{metaAndPlot}
\title{Perform GLMM or inverse variance meta-analysis}
\usage{
metaAndPlot(
  lst,
  name.geneset,
  nperm = 500,
  nboot = 200,
  method = c("GLMM", "Inverse")[1],
  effect = c("random", "fixed")[1]
)
}
\arguments{
\item{lst}{list of GSAR_boot() results for multiple datasets regarding one gene set}

\item{name.geneset}{name of the concerned single gene set}

\item{nperm}{times of sample indix permutation, necessitated by GSNCA}

\item{nboot}{number of bootstrap times}

\item{method}{method for meta analysis, either GLMM or Inverse (inverse)}

\item{effect}{choice of effects model, either fixed or random}
}
\value{
metaprop function result object together with p-values for multiple bootstrap meta-analyses.
}
\description{
Perform GLMM or inverse variance meta-analysis on original and bootstrap analysis results
}
\details{
1.GLMM may reject to apply if all individual datasets have identical proportion, so there is always special treatment on GLMM track considering this marginal condition. For bootstrap sample meta-analysis, used tryCatch() to ignore occasional marginal situation on individual bootstrap time. 2. Forest plot is generated in a PDF file. 3. method designates GLMM or Inverse. If inappropriate, change method to method.choice throughout. 4. Use single generic term to designate either fixed or random series.
}
