% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetaGSCA.R
\name{MetaGSCA}
\alias{MetaGSCA}
\title{Meta-analysis of differential coexpression analysis}
\usage{
MetaGSCA(
  list.geneset,
  list.dataset,
  list.group,
  names.geneset = names(list.geneset),
  names.dataset = names(list.dataset),
  nperm = 500,
  nboot = 200,
  method = c("GLMM", "Inverse")[1],
  effect = c("random", "fixed")[1]
)
}
\arguments{
\item{list.geneset}{a list of gene sets (one or multiple).}

\item{list.dataset}{a list of datasets, first column is gene name.}

\item{list.group}{a list of samples/patients subgroup or condition (e.g. (1,1,1,2,2,2)).}

\item{names.geneset}{gene set names, corresponding to list.geneset}

\item{names.dataset}{dataset names, corresponding to list.dataset}

\item{nperm}{number of permutations used to estimate the null distribution of the test statistic. If not given, a default value 500 is used.}

\item{nboot}{number of bootstraps used to estimate the point and interval estimate. If not given, a default value 200 is used.}

\item{method}{meta-analysis method. Must be either `GLMM` or `Inverse`.}

\item{effect}{statistical model applied in meta-analysis. Must be either `random` or `fixed`.}
}
\description{
Meta-analysis of differential coexpression analysis across multiple datasets on multiple gene sets
}
\details{
`MetaGSCA` systematically assesses the coexpression disturbance of a gene set by pooling the results from individual studies. In the kernel, a nonparametric approach named GSNCA tests whether a gene set is differentially coexpressed between two comparative conditions, and it produces a set-wise coexpression distance statistics as well as a permutation-based p-statistics. A meta-analysis is then performed to combine individual study results with either of two options: a random-intercept logistic regression model and the inverse variance method. The current function can take as input both a single gene set and multiple gene sets.
In the meta-analysis of proportion method, we could perform fixed- and random- effects model. If we choose the fixed-effects model, we assume that the parameter of interest is identical across studies and any difference between the observed proportion is only due to sampling error. If we choose the random-effects model, we assume that the observed difference between the proportions cannot be entirely attributed to sampling error and other factors such as differences in study popula-tion, study designs, etc. To examine con-sistency of findings across studies, a statistical test of heterogeneity is reported in the meta-analysis. If heterogeneity is a concern, the random-effects model is recommended. In this case, each study estimates a different parameter, and the pooled estimate describes the mean proportion of the distribution. The variance parameter de-scribes the heterogeneity among the studies (when the variance is zero, this is equivalent to the fixed-effects model).
Logit transformation is applied to the binary outcome and regressed on the study variable in fixed effects or random effects logistic model. Define the logit(p) as effect size (ES), it can be estimated using two approaches: in-verse variance method and generalized linear mixed model (GLMM) with fixed intercept. The interval estimation is provided from these methods as well as from the bootstrap approach.
}
\examples{
data(meta)
data3 <- c('BRCA','COAD','HNSC')
testSingle <- MetaGSCA(list.geneset = genesets[2],
        list.dataset = datasets[data3], 
        list.group = groups[data3],
        names.geneset = names(genesets)[2],
        names.dataset = data3,
        nperm = 100,
        nboot = 100,
        method = 'GLMM',
        effect = 'random')

}
