% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetaGSCA_plot.R
\name{MetaGSCA_plot}
\alias{MetaGSCA_plot}
\title{Draw Forest plot based on metaprop() result}
\usage{
MetaGSCA_plot(meta, meta.p, distinct.label)
}
\arguments{
\item{meta}{output object of metaprop function on original set of datasets}

\item{meta.p}{p-value vector obtained by metaprop on bootstrap series of datasets}

\item{distinct.label}{bottom text line on forestplot with nperm, nboot, and p summary}
}
\description{
Draw Forest plot based on meta result and meta-p vectors on bootstrap runs
}
