% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PWcTalk.R
\name{PWcTalk}
\alias{PWcTalk}
\title{Streamlined functions for Pathway Crosstalk Network analysis}
\usage{
PWcTalk(
  input2PWcTalk,
  test = "binary",
  pTh.dataset = 0.01,
  pTh.pwPair = 0.01,
  pTh.pw = 0.01,
  figname = "PWcTalk",
  pdfW = 10,
  pdfH = 10,
  asp = 0.7,
  vbase = 15,
  ebase = 2,
  vlbase = 1,
  power = 1/2
)
}
\arguments{
\item{input2PWcTalk}{A data frame with 330 observations across 14 variables. Each observation corresponds to a pathway}

\item{test}{Similarity measure over two binary vectors, either asymmetric Binary similarity or Pearson's Phi index}

\item{pTh.dataset}{Threshold for permutation p-statistic of a gene set co-expression in a dataset}

\item{pTh.pwPair}{Threshold for the similarity measure between two pathways}

\item{pTh.pw}{Threshold for the meta-analysis p-statistic of a pathway’s co-expression across multiple datasets}

\item{figname}{Stem file name for the resultant pathway crosstalk network figure}

\item{pdfW}{Width of PDF figure file}

\item{pdfH}{Height of PDF figure file}

\item{asp}{Aspect ratio (between width and height). Asp parameter of function plot.igraph}

\item{vbase}{A base value regarding vertex size. Vertex size depends on vbase and a soft thresholding Power parameter}

\item{ebase}{A base value regarding edge width. Edge width depends on ebase and a soft thresholding Power parameter}

\item{vlbase}{A base value regarding verbex label. Vertex label font depends on vlbase and a soft thresholding Power parameter}

\item{power}{A power paramter involved in soft-thresholding treatment. Power is implicated in vertex size, edge width, and vertex label font}
}
\value{
An igraph graph (network) object
}
\description{
Overall main function packed as a cascade of major sub-functions
}
\details{
Nevertheless, it is recommended to collapse the function and run the sub-functions in a script instead.
}
\examples{
data(input2PWcTalk)
# One code line to execute pathway crosstalk analysis. Simple, but no flexibility for layout tuning. 
PWcTalk(input2PWcTalk,test='binary',
 pTh.dataset=0.01,pTh.pwPair=0.01,pTh.pw=0.01,figname='PWcTalk',
 pdfW=10,pdfH=10,asp=0.7,vbase=15,ebase=2,vlbase=1,power=1/2)
}
\seealso{
[PWcTalkNWpre()] for the first module of prior steps, [PWcTalkNW()] for the second module of post steps.

Other aggregations: 
\code{\link{PWcTalkNWpre}()},
\code{\link{PWcTalkNW}()}
}
\concept{aggregations}
