% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PWsimDichot.R
\name{PWsimDichot}
\alias{PWsimDichot}
\title{Dichotomize p-values of PW pairs according to pair p threshold (say 0.01) [and vertex metaP threshold (say 0.01)]}
\usage{
PWsimDichot(pwPairP, metaP = NULL, pTh.pwPair = 0.01, pTh.pw = 0.01)
}
\arguments{
\item{pwPairP}{output from function PWsim}

\item{metaP}{list component $metaP of function dichotGSAR output}
}
\description{
Dichotomize p-values of PW pairs according to pair p threshold (say 0.01) [and vertex metaP threshold (say 0.01)]
}
